# hardware_store/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для магазина бытовой техники",
        "title": "Hardware Store API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает список всех продуктов в системе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить список продуктов",
                "responses": {
                    "200": {
                        "description": "Список продуктов успешно получен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hardware_store_internal_web_dto.ProductResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении списка",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Создаёт новый продукт в системе на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Создать новый продукт",
                "parameters": [
                    {
                        "description": "Данные продукта для создания",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Продукт успешно создан",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибки валидации полей или некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при сохранении продукта",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает полные данные продукта по уникальному идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить продукт",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Продукт успешно получен",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении продукта",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Удаляет продукт по уникальному идентификатору UUID",
                "tags": [
                    "products"
                ],
                "summary": "Удалить продукт",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Продукт успешно удалён"
                    },
                    "400": {
                        "description": "Невалидный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при удалении",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/products/{id}/stock": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Обновляет доступное количество товара на складе для указанного продукта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Обновить количество товара на складе",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое количество товара",
                        "name": "stock",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.UpdateStockCountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Количество товара успешно обновлено",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный формат запроса, отрицательное количество или недостаточный остаток",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Продукт не найден",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при обновлении",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "hardware_store_internal_web_dto.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "hardware_store_internal_web_dto.NotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "client not found"
                }
            }
        },
        "hardware_store_internal_web_dto.ProductRequest": {
            "description": "Запрос на обновление количества товара на складе",
            "type": "object",
            "required": [
                "available_stock",
                "category_id",
                "name",
                "price",
                "supplier_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 15
                },
                "category_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Холодильник Samsung RB38A7861B1"
                },
                "price": {
                    "type": "number",
                    "example": 75990
                },
                "supplier_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "hardware_store_internal_web_dto.ProductResponse": {
            "description": "Полная информация о товаре включая цену, остатки и информацию о поставщике",
            "type": "object",
            "required": [
                "available_stock",
                "last_update_date",
                "name",
                "price",
                "product_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "category": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "image": {
                    "type": "string",
                    "example": "b1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
                },
                "last_update_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string",
                    "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
                },
                "supplier": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "hardware_store_internal_web_dto.UpdateStockCountRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "hardware_store_internal_web_dto.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid uuid format"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: \"Bearer \u003cваш_токен\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
                "email": "Просмотр вашего email",
                "profile": "Просмотр вашего профиля"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        },
        {
            "OAuth2": []
        }
    ]
}